streamlit/requirements.txt
requests
streamlit==0.58.0
pandas
mysql-connector-python
sqlalchemy
pymysql